Fixes/improvements to PSD model ModelFDDataPSD   -Fixed spurious offset in fit for log-scale fits   -Added option 'logratio' for fits based on log of ratio to model   -Reorg of code to better support the increase number of options   -Adjusted the way spline knots are located to also fill in at low-f    when the number of knots is increased.   -More strenuous conditions applied to knows to better avoid errors    with larger numbers of knots.   -Moved the code from connect/fillgaps.py to psdmodel.py   -New notebook including just the development of this model Spritz-PSD-modeling
