+++ bayesdawn/connect/fillgaps.py
@@ -826,265 +826,3 @@ def LDC_imputation(data_masked, maskinfo, psd_correction, names = ['A', 'E', 'T'
         return data_rec
 
 
-#The PSD model class
-class ModelFDDataPSD(psdmodel.PSD):
-    '''
-    Specialization of the bayesdawn psd model class which provides a bayesdawn PSD model 
-    build on a possible combination of analytic (LDC) noise models possibly with smoothing 
-    and/or additional data fitting.
-    
-    Parameters
-    ----------
-    data : numpy rec-array (or dict)
-        Multichannel Fourier data set [named columns 'f','A','E',...
-    channel : str
-        Tag specifying the channel to use for this instance, eg 'A', 'E'
-    fit_type :  str = 'poly' 
-        Tag for the type of model fitting to apply, or None to use the LDC model.
-        Supported ['poly', 'log_poly', None]
-    fit_dof : int [default 4]
-        Number of degrees of freedom in the fit.
-    fit_logx : bool [default True]
-        If True use log(f) as the coordinate for fitting functions.
-    noise_model : str [default 'spritz']
-        Tag for the LDC base model to use for ratio, or None for no ratio.
-    smooth_df: float [default None]
-        Freq width scale to set number of points for smoothing
-    fmin: float [default 1e-5]
-        Minimum frequency for result
-    fmax: float [default None]
-        Maximum frequency for result
-        
-        
-    Returns
-    -------
-    bayesdawn.psfmodel.PSD object
-    
-    For the specified channel analytic PSD model will be generated. If noise_model is not None,
-    then the model will be generated by a ratio against the specified analytic noise model.
-    
-    If smooth_df is given, then the analytic noise model will be smoothed, by convolution with a
-    Hanning window. 
-    
-    The type of fitting is specifed by fit_type, with 'poly' providing a fit using np.polyfit with 
-    polynomial degree fit_dof-1. A variation on this is 'log_poly' which fits the ratio of the log 
-    of the two quantities, scaled by the sum of their max values. Other methods may be added. 
-    
-    If fit_type is None, then the analytic model (possibly smoothed) is applied directly and only 'f' 
-    will be used from the data.
-    
-    Note 
-    ----
-    
-    Long-term, it probably makes more sense to fold the functionality here into the underlying
-    bayesdawn.psdmodel.PSD code, which already includes a MCMC sampled spline fit for the PSD. Missing
-    there is the crutch we use here for starting from, and scaling against an approximate analytic model.
-    '''
-
-    def __init__(self, data, channel, fit_type='poly',fit_dof=4, fit_logx=True, noise_model='spritz', smooth_df=None, fmin=1e-5, fmax=None):
-    
-        self.channel = channel
-        self.fit_type = fit_type
-        self.chdata=data[channel]
-
-        f=data['f']
-        df=(f[-1]-f[0])/(len(f)-1) 
-        self.df=df
-        
-        fs=f[-1]*2
-        ndata=(len(f)-1)*2
-        self.ndata=ndata
-        
-        if fmax is not None:
-            self.chdata=self.chdata[f<=fmax]
-            f = f[f<=fmax]
-        if fmin is not None:
-            self.chdata=self.chdata[f>=fmin]
-            f = f[f>=fmin]
-            
-        self.fin=f.copy()
-        
-        psdmodel.PSD.__init__(self, ndata, fs, fmin=fmin, fmax=fmax)
-        
-        self.scalefac=2*self.df
-        
-        if noise_model is not None:
-            Nmodel=get_noise_model(noise_model, f)
-            Sinit = Nmodel.psd(tdi2=True, option=self.channel, freq=f)
-            if smooth_df is not None:
-                nsmooth=int(smooth_df/df)
-                w=np.hanning(nsmooth)
-                smooth=lambda s:np.convolve(w/w.sum(),s,mode='same')
-                Sinit=smooth(Sinit)
-        else: 
-            Sinit=None
-            if fit_type is None: raise ValueError("Must specify at least fit_type or noise_model")
-        self.Sinit=None
-        if Sinit is not None:
-            self.logSinit=scipy.interpolate.interp1d(np.log(f),np.log(Sinit),fill_value="extrapolate")
-            self.Sinit=lambda x:np.exp(self.logSinit(np.log(x)))
-        
-        self.fit=None
-        if fit_type is None: return
-        self.fit_dof=fit_dof
-        
-        self.fit_logx=fit_logx
-        if fit_logx: x=np.log(f)
-        else: x=f
-    
-        if not fit_type.startswith('log_'):
-            print('Fitting on linear-scale')
-            if fit_type=='poly' or fit_type=='spline':
-                y=np.abs(self.chdata)**2*self.scalefac
-                
-                if self.Sinit is not None:
-                    #print('Sinit min/max',min(Sinit),max(Sinit))
-                    y=y/self.Sinit
-        else:
-            print('Fitting on log-scale')
-            if fit_type=='log_poly' or fit_type=='log_spline':
-                y=np.abs(self.chdata)**2*self.scalefac
-                if self.Sinit is not None:
-                    zero=2*(max(y)+max(Sinit))
-                    #print('Sinit min/max',min(Sinit),max(Sinit))
-                    y=np.log(y/zero)/np.log(self.Sinit(self.fin)/zero)
-                else:
-                    zero=2*(max(y))
-                    y=np.log(y/zero)
-                self.fit_zero=zero
-                
-        if fit_type=='poly' or fit_type=='log_poly':              
-            pf = np.polyfit(x,y,fit_dof+1,w=1/f)
-            #print('poly fit:',pf)
-            self.fit = np.poly1d(pf)
-        elif fit_type=='spline' or fit_type=='log_spline':
-            # I tried using the functionality in psdmodel.py but couldn't get it working.
-            # to stay close to the existing implementation in psdmodel.py                    
-            print(f[0],'< f < ',f[-1],'fmin/fmax=',fmin,fmax)
-            n_knots=fit_dof-2 #dof=n_knots+2 (maybe, sort of guessing)                    
-            knots=self.choose_knots(f) #This function needs f, not x
-            xknots=knots
-            if fit_logx: xknots=np.log(knots) #transform to x-space if needed
-            fitspline=scipy.interpolate.LSQUnivariateSpline(x, y, xknots[1:-1], w=1/f, k=3, ext=3, check_finite=False)                    
-            print('knots',knots)
-            self.knots=knots
-            self.fit=fitspline
-        else:    
-            raise ValueError('fit_type '+str(fit_type)+' not recognized.')
-            
-
-            
-    def choose_knots(self,x):
-        '''
-        Arguments:
-                  x : Numpy array
-                      Data coordinates for which the spline will be applied. These are used for setting the 
-                      spline range and for checking the suitability of the knots.
-                 
-        Returns:
-             Numpy array with the ordered knots
-             
-        For spline data we want to set the location of the knots. There are diverse ways to do this
-        of course.  Here we follow the scheme in bayesdawn.psdmodel (fixed).  For B-splines to work
-        the knots have to meet a constraint, dependent on the data.  From the scipy documentation
-        for LSQUnivariateSpline:
-           "Knots t must satisfy the Schoenberg-Whitney conditions, i.e., there must be a subset of 
-            data points x[j] such that t[j] < x[j] < t[j+k+1], for j=0, 1,...,n-k-2."
-        If the conditions aren't met, knots will be dropped until the condition is met, this may result
-        in an actual number of knots less than desired to realize the value of fit_dof. 
-        [We could add a interative loop to try to improve that.]
-        '''
-        print('Finding knots in ',x[0],'< x <',x[-1])
-        nknots=self.fit_dof #Approx guess of the relation of knots to dof
-        t=psdmodel.choose_frequency_knots(nknots+2, freq_min=x[0], freq_max=x[-1])
-        t=t[1:-1] #drop the boundaries (scipy will put them in)
-        print(len(t),'knots before checking')
-        
-        #Now we check the Schoenberg-Whitney condition
-        j=0
-        k=3
-        iscan=0
-        while j+k+1<len(t):
-            #First scan to the relevant part of the data
-            while x[iscan]<=t[j] and iscan<len(x):iscan+=1
-            #Now look for a data point in the window
-            i=iscan
-            #check whether the next point exists and is in the target range
-            if iscan>=len(x) or x[iscan]>=t[j+k+1]: 
-                #fails the condition. To fix, we remove an intermediate knot and try again
-                jkill=j+k//2+1
-                print('Dropping knot at ',t[jkill],'because no data in',t[j],'< t <',t[j+k+1])
-                t=np.delete(t,jkill)
-            else:
-                j+=1
-        print(len(t),'knots after ensuring condition.')
-        return t
-    
-    def plot(self,show_fit=False, tag=None):
-        
-        if tag is None: 
-            tag=self.fit_type
-        if tag is None:
-            tag='no fit'
-            
-        plt.loglog(self.fin,np.abs(self.chdata*np.sqrt(self.scalefac)),label='data ('+tag+')',alpha=0.3)
-        plt.loglog(self.fin,np.sqrt(self.psd_fn(self.fin)),label='model ('+tag+')')
-        plt.legend()                
-            
-        if show_fit and self.fit is not None and self.Sinit is not None:
-            if self.fit_logx: x=np.log(self.fin)
-            else: x=self.fin
-            
-            y=np.abs(self.chdata)**2*self.scalefac
-            if not self.fit_type.startswith('log_'):
-                if self.Sinit is not None:
-                    #print('Sinit min/max',min(Sinit),max(Sinit))
-                    y=y/self.Sinit(self.fin)
-            else:
-                zero=self.fit_zero
-                y=np.log(y/zero)
-                if self.Sinit is not None:
-                    y=y/(self.logSinit(np.log(self.fin))-np.log(zero))
-                    
-            print('x',x[:5])
-            print('y',y[:5])
-            print('fit',y[:5])
-                
-            spline=self.fit_type.endswith('spline')
-                
-            plt.show()
-            plt.semilogy(self.fin,(y),label='data ratio ('+tag+')')
-            plt.semilogy(self.fin,self.fit(x),label='fit ('+tag+')')
-            if spline:
-                xknots=self.knots
-                if self.fit_logx: xknots=np.log(xknots)
-                
-                plt.semilogy(self.knots,self.fit(xknots),'*',label='knots ('+tag+')')
-            plt.legend()
-            plt.show() 
-            if False:
-                plt.loglog(self.fin,(y),label='data ratio ('+tag+')')
-                plt.loglog(self.fin,self.fit(x),label='fit ('+tag+')')
-                plt.legend()
-                plt.show() 
-
-    def psd_fn(self, x):
-        # returns the psd function defined earlier           
-
-        if self.fit is not None:
-            xx=x.copy()
-            if self.fit_logx:
-                xx=np.log(x)
-            if not self.fit_type.startswith('log_'):
-                dm = np.abs(self.fit(xx))            
-                if self.Sinit is not None:
-                    dm = dm*self.Sinit(x)
-            else:
-                logdm = self.fit(xx)
-                zero=self.fit_zero
-                if self.Sinit is not None:
-                    logdm*=(self.logSinit(np.log(x))-np.log(zero))
-                dm=zero*np.exp(logdm)
-        else:
-            dm = self.Sinit(x)
-        return dm
