+++ bayesdawn/psdmodel.py
@@ -243,12 +243,12 @@ def choose_frequency_knots(n_knots, freq_min=1e-5, freq_max=1.0, base=10):
     targetfunc = lambda x: n0 - (1 - x ** (n_knots)) / (1 - x) - ns
     result = optimize.fsolve(targetfunc, alpha_guess)
     alpha = result[0]
-    print('alpha',alpha)
+    #print('alpha',alpha)
     n_knots = ns + (1 - alpha ** jvect) / (1 - alpha)
     f_knots = base ** (n_knots)
-    print('knots before trim:',f_knots)
+    #print('knots before trim:',f_knots)
     f_knots = f_knots[(f_knots >= freq_min) & (f_knots <= freq_max)]
-    print(len(f_knots))
+    #print(len(f_knots))
     
     return np.unique(np.sort(f_knots))
 
@@ -1179,3 +1179,424 @@ class PSDTheoretical(PSD):
         logp_values_list = [samp[1] for samp in sampling_result]
 
         return sample_list, logp_values_list
+
+
+
+
+
+
+
+# =============================================================================
+# Multipurpose PSD model
+# =============================================================================
+class ModelFDDataPSD(PSD):
+    '''
+    Specialization of the bayesdawn psd model class which provides a bayesdawn PSD model 
+    build on a possible combination of analytic (LDC) noise models possibly with smoothing 
+    and/or additional data fitting.  
+    
+    Parameters
+    ----------
+    data : numpy rec-array (or dict)
+        Multichannel Fourier data set [named columns 'f','A','E',...
+    channel : str
+        Tag specifying the channel to use for this instance, eg 'A', 'E'
+    fit_type :  str = 'poly' 
+        Tag for the type of model fitting to apply, or None to use the LDC model.
+        Supported ['poly', 'log_poly', 'spline', ..., 'logratio_spline']
+    fit_dof : int [default 4]
+        Number of degrees of freedom in the fit.
+    fit_logx : bool [default True]
+        If True use log(f) as the coordinate for fitting functions.
+    noise_model : str [default 'spritz']
+        Tag for the LDC base model to use for ratio, or None for no ratio.
+    smooth_df: float [default None]
+        Freq width scale to set number of points for smoothing
+    fmin: float [default 1e-5]
+        Minimum frequency for result
+    fmax: float [default None]
+        Maximum frequency for result
+    offset_log_fit: bool [default True]
+        Correct the PSD bias caused by fitting in log space. Only set this to False to demonstrate the bias.
+        
+    Returns
+    -------
+    bayesdawn.psfmodel.PSD object
+    
+    For the specified channel analytic PSD model will be generated. If noise_model is not None,
+    then the model will be generated by a ratio against the specified analytic noise model.
+    
+    If smooth_df is given, then the analytic noise model will be smoothed, by convolution with a
+    Hanning window. 
+    
+    The type of fitting is specifed by fit_type, with 'poly' providing a fit using np.polyfit with 
+    polynomial degree fit_dof-1. A variation on this is 'log_poly' which fits the ratio of the log 
+    of the two quantities, scaled by the sum of their max values. The 'logratio_poly' option fits
+    the log of the ratio vs the model function.  The 'spline', 'log_spline' and 'logratio_spline' 
+    options work analogously, but fit the data to a B-spline, with the number of knots guided by the
+    fit_dof.  
+    
+    If fit_type is None, then the analytic model (possibly smoothed) is applied directly and only 'f' 
+    will be used from the data.
+
+    
+    
+    Notes 
+    ----
+    
+    This derived class doesn't yer support a MCMC sampled spline fit for the PSD, though that should
+    not be too difficult.
+
+    This class assumes, and builds off Fourier data, though it would not be difficult to also allow
+    time-domain data directly and then to compute an FT for it.
+    
+    '''
+
+    def __init__(self, data, channel, fit_type='logratio_spline',fit_dof=10, fit_logx=True, noise_model='spritz', smooth_df=None, fmin=1e-5, fmax=None, offset_log_fit=True):
+    
+        self.channel = channel
+        self.fit_type = fit_type
+        self.chdata=data[channel]
+
+        f=data['f']
+        df=(f[-1]-f[0])/(len(f)-1) 
+        self.df=df
+        
+        fs=f[-1]*2
+        ndata=(len(f)-1)*2
+        self.ndata=ndata
+        
+        if fmax is not None:
+            self.chdata=self.chdata[f<=fmax]
+            f = f[f<=fmax]
+        if fmin is not None:
+            self.chdata=self.chdata[f>=fmin]
+            f = f[f>=fmin]
+            
+        self.fin=f.copy()
+        
+        PSD.__init__(self, ndata, fs, fmin=fmin, fmax=fmax)
+        
+        self.scalefac=2*self.df
+        
+        if noise_model is not None:
+            from ldc.lisa.noise import get_noise_model
+            Nmodel=get_noise_model(noise_model, f)
+            Sinit = Nmodel.psd(tdi2=True, option=self.channel, freq=f)
+            if smooth_df is not None:
+                nsmooth=int(smooth_df/df)
+                w=np.hanning(nsmooth)
+                smooth=lambda s:np.convolve(w/w.sum(),s,mode='same')
+                Sinit=smooth(Sinit)
+        else: 
+            Sinit=None
+            if fit_type is None: raise ValueError("Must specify at least fit_type or noise_model")
+        self.Sinit=None
+        if Sinit is not None:
+            self.logSinit=interpolate.interp1d(np.log(f),np.log(Sinit),fill_value="extrapolate")
+            self.Sinit=lambda x:np.exp(self.logSinit(np.log(x)))
+        
+        self.fit=None
+        if fit_type is None: return
+        self.fit_dof=fit_dof
+        
+        self.fit_logx=fit_logx
+        if fit_logx: x=np.log(f)
+        else: x=f
+
+        # We internally hold info about the fit type in the combination of
+        #     fit_scale:  in ['linear', 'log', 'logratio']
+        #  and fit_func:  in ['poly', 'spline']
+        if not fit_type.startswith('log'):
+            fit_scale='linear'
+            fit_func=fit_type
+            #print('Fitting on linear-scale')
+        elif fit_type.startswith('log_'):
+            fit_scale='log'
+            fit_func=fit_type[4:]
+            #print('Fitting on log-scale')
+        elif fit_type.startswith('logratio_'):
+            fit_scale='logratio'
+            fit_func=fit_type[9:]
+            #print('Fitting on log-scale')
+        else: raise ValueError('Fit scale not understood for fit_type'+fit_type)
+        # Now a check:
+        if not fit_func in ['poly','spline']: raise ValueError("Couldn't process fit_type '"+fit_type+"', fit_func='"+fit_func+"'")        
+        self.fit_scale=fit_scale
+        self.fit_func=fit_func
+
+        # prepare the data for fitting
+        y=np.abs(self.chdata)**2*self.scalefac                
+        if not fit_scale=='log' and self.Sinit is not None:
+            zero=2*(max(y)+max(Sinit))
+        else: zero=1
+        self.fit_zero=zero
+
+        if self.Sinit is not None:
+            if fit_scale=='linear':
+                y=y/self.Sinit(self.fin)
+            elif fit_scale=='log':
+                y=np.log(y/zero)/np.log(self.Sinit(self.fin)/zero)
+            elif fit_scale=='logratio':
+                y=np.log(y/self.Sinit(self.fin))
+            else: raise ValueError('Unknown fit_scale '+fit_scale)
+        else:
+            if fit_scale=='linear':
+                pass
+            elif fit_scale=='log':
+                y=np.log(y/zero)
+            elif fit_scale=='logratio':
+                y=np.log(y)
+            else: raise ValueError('Unknown fit_scale '+fit_scale)
+
+        # Offset the data when the fit is in log space
+        # Because the mean of log is lower than the log of the mean by euler_gamma
+        self.offset_log_fit = offset_log_fit and fit_scale in ['log','logratio']
+        
+        #perform the fit
+        if fit_func=='poly':
+            pf = np.polyfit(x,y,fit_dof+1,w=1/f)
+            #print('poly fit:',pf)
+            fitpoly=np.poly1d(pf)
+            self.fit = lambda x: fitpoly(x)
+        elif fit_func=='spline':
+            # I tried using the functionality in psdmodel.py but couldn't get it working.
+            # to stay close to the existing implementation in psdmodel.py                    
+            print(f[0],'< f < ',f[-1],'fmin/fmax=',fmin,fmax)
+            n_knots=fit_dof-2 #dof=n_knots+2 (maybe, sort of guessing)                    
+            knots=self.choose_knots(f) #This function needs f, not x
+            xknots=knots
+            print('lowest f knots:',knots[:5])
+            if fit_logx: xknots=np.log(knots) #transform to x-space if needed
+            try:
+                fitspline=interpolate.LSQUnivariateSpline(x, y, xknots[1:-1], w=1/f, k=3, ext=3, check_finite=False)
+            except ValueError:
+                
+                print('Problem creating spline:')
+                t= xknots[1:-1]
+                print('inputs:\n  x=',x,'\n  y=',y,'\n  t=', t,'\n  w=',1/f)
+                # _data == x,y,w,xb,xe,k,s,n,t,c,fp,fpint,nrdata,ier
+                xb = x[0]
+                xe = x[-1]
+                k=3
+                t = np.concatenate(([xb]*(k+1), t, [xe]*(k+1)))
+                n = len(t)
+                print('internal form of t',t)
+                if not np.all(t[k+1:n-k]-t[k:n-k-1] > 0, axis=0):
+                    raise ValueError('Interior knots t must satisfy '
+                                     'Schoenberg-Whitney conditions')
+                print('1) k+1 <= n-k-1 <= m:',k+1,n-k-1,len(x))
+                print('2a) t(1) <= t(2) <= ... <= t(k+1)',t[:k+1])
+                print('2b) t(n-k) <= t(n-k+1) <= ... <= t(n)',t[n-k-1:])
+                print('3) t(k+1) < t(k+2) < ... < t(n-k): min(t[x+1]-t[x]) over interior range:',min(t[k+1:n-k]-t[k:n-k-1]))
+                print('4) t(k+1) <= x(i) <= t(n-k): t[k],min(x),max(x),t[n-k-1]', t[k],min(x),max(x),t[n-k-1])
+                print('5): Schoenberg and Whitney condition')
+                #Now we check the Schoenberg-Whitney condition
+                j=0
+                k=3
+                iscan=0
+                while j+k+1<len(t):
+                    nmin=1#minimum number of points in the segment
+                    #First scan to the relevant part of the data
+                    while x[iscan]<=t[j] and iscan<len(x):iscan+=1
+                    #check whether the next point exists and is in the target range
+                    if iscan>=len(x) or x[iscan]>=t[j+k+1]: 
+                        #fails the condition. To fix, we remove an intermediate knot and try again
+                        jkill=j+k//2+1
+                        print('Fails for knot at ',t[jkill],'because no data in',t[j],'< t <',t[j+k+1])
+                        print('nearby data:',x[iscan-2:iscan+4])
+                    j+=1
+                
+                import scipy.interpolate.dfitpack as dfitpack
+                if not dfitpack.fpchec(x, t, k) == 0:
+                    print('Failed')
+                scipy.interpolate.LSQUnivariateSpline(x, y, xknots[1:-1], w=1/f, k=3, ext=3, check_finite=False)
+                    
+            self.knots=knots
+            self.fit=lambda x:fitspline(x)
+                
+        else:    
+            raise ValueError('fit_func '+str(fit_func)+' not recognized.')
+            
+
+            
+    def choose_knots(self,x,verbose=False):
+        '''
+        Arguments:
+                  x : Numpy array
+                      Data coordinates for which the spline will be applied. These are used for setting the 
+                      spline range and for checking the suitability of the knots.
+                 
+        Returns:
+             Numpy array with the ordered knots
+             
+        For spline data we want to set the location of the knots. There are diverse ways to do this
+        of course.  Here we follow the scheme in bayesdawn.psdmodel (fixed).  For B-splines to work
+        the knots have to meet a constraint, dependent on the data.  From the scipy documentation
+        for LSQUnivariateSpline:
+           "Knots t must satisfy the Schoenberg-Whitney conditions, i.e., there must be a subset of 
+            data points x[j] such that t[j] < x[j] < t[j+k+1], for j=0, 1,...,n-k-2."
+        If the conditions aren't met, knots will be dropped until the condition is met, this may result
+        in an actual number of knots less than desired to realize the value of fit_dof. 
+        [We could add a interative loop to try to improve that.]
+        '''
+        if verbose: print('Finding knots in ',x[0],'< x <',x[-1])
+        nknots=self.fit_dof #Approx guess of the relation of knots to dof
+        minf=x[0]
+        maxf=x[-1]
+        base=(maxf/minf)**(1/nknots)
+        print('base=',base)
+        # We use this choose_frequency_knots function, decreasing the 'base' value when
+        # there are more knots desired to ensure that the benefit of more knots also
+        # shows at lower frequencies. It is not clear whether this treatment is any
+        # better than just even logarithmic sampling.
+        t=choose_frequency_knots(nknots+2, freq_min=x[0], freq_max=x[-1],base=base)
+        t=np.concatenate(([x[0]],t,[x[-1]])) #we add on points at the boundary to ensure that the check condition check is thorough
+        if verbose: print(len(t),'knots before checking')
+        i=0
+        tlast=x[0]
+        while i<len(t):
+            if t[i]<=tlast: #not strictly ordered
+                if verbose: print('Dropping knot at ',t[i],'because not beyond last knot at',tlast)
+                t=np.delete(t,i)
+            else:
+                tlast=t[i]
+                i+=1
+                
+        
+        #Now we check the Schoenberg-Whitney condition
+        j=0
+        k=1 #we use 1 instead of 3 to be conservative, and to hopefully avoid issues that seem to happen near ends
+        iscan=0
+        while j+k+1<len(t):
+            nmin=1#minimum number of points in the segment
+            #First scan to the relevant part of the data
+            while x[iscan]<=t[j] and iscan<len(x):iscan+=1
+            #check whether the next point exists and is in the target range
+            if iscan>=len(x) or x[iscan]>=t[j+k+1]: 
+            #if iscan>=len(x) or x[iscan]>=t[j+k]: #Artificially tighten condition
+                #fails the condition. To fix, we remove an intermediate knot and try again
+                jkill=j+k//2+1
+                if verbose: print('Dropping knot at ',t[jkill],'because no data in',t[j],'< t <',t[j+k+1])
+                t=np.delete(t,jkill)
+            else:
+                j+=1
+        if verbose: print(len(t),'knots after ensuring condition.')
+        t=t[1:-1] #drop the boundaries (scipy will put them in)
+        return t
+    
+    def plot(self,show_fit=False, tag=None, ref=None):
+        '''
+        Interactively how a plot illustrating how the PSD model compares with the data.
+        Arguments:
+           show_fit: bool
+             Include a second plot illustrating directly how the fit (if any) works
+           tag: string
+             A label to include in the plot title
+           ref: psdmodel
+             Another PSD model to serve as a reference in the plot
+        '''        
+        
+        if tag is None: 
+            tag=self.fit_type
+        if tag is None:
+            tag='no fit'
+
+        import matplotlib.pyplot as plt
+        plt.loglog(self.fin,np.abs(self.chdata*np.sqrt(self.scalefac)),label='data ',alpha=0.3)
+        if ref is not None:
+            plt.loglog(self.fin,np.sqrt(ref.psd_fn(self.fin)),label='ref')
+        plt.loglog(self.fin,np.sqrt(self.psd_fn(self.fin)),label='model ('+tag+')')
+            
+        plt.legend()                
+            
+        if show_fit and self.fit is not None and self.Sinit is not None:
+            if self.fit_logx: x=np.log(self.fin)
+            else: x=self.fin
+
+            if ref is not None:
+                y0=ref.psd_fn(self.fin)
+            else:
+                y0=1
+                
+            y=np.abs(self.chdata)**2*self.scalefac
+            fit_scale=self.fit_scale
+            if fit_scale=='linear':
+                if self.Sinit is not None:
+                    #print('Sinit min/max',min(Sinit),max(Sinit))
+                    fac=self.Sinit(self.fin)
+                    y=y/fac
+                    y0=y0/fac
+            elif fit_scale=='log':
+                zero=self.fit_zero
+                y=np.log(y/zero)
+                y0=np.log(y0/zero)
+                if self.Sinit is not None:
+                    fac=(self.logSinit(np.log(self.fin))-np.log(zero))
+                    y=y/fac
+                    y0=y0/fac
+            elif fit_scale=='logratio':
+                y=np.log(y)
+                y0=np.log(y0)
+                if self.Sinit is not None:
+                    #off=self.logSinit(np.log(self.fin))
+                    off=np.log(self.Sinit(self.fin))
+                    y=y-off
+                    y0=y0-off
+            else: raise ValueError("fit_scale unknown")
+
+            spline=self.fit_func=='spline'
+            if False:
+                print('x',x[:5])
+                print('y',y[:5])
+                print('fit',y[:5])
+                print('y v y0 min,max:',min(y),min(y0),max(y),max(y0))
+
+            plt.show()
+
+            pltfunc=plt.plot
+            pltfunc(self.fin,(y),label='data ratio ')
+            pltfunc(self.fin,(y0),label='ref ratio ')
+            pltfunc(self.fin,self.fit(x),label='fit ('+tag+')')
+            if spline:
+                xknots=self.knots
+                if self.fit_logx: xknots=np.log(xknots)
+                
+                pltfunc(self.knots,self.fit(xknots),'*',label='knots ('+tag+')')
+            plt.legend()
+            plt.show() 
+            if False:
+                plt.loglog(self.fin,(y),label='data ratio ('+tag+')')
+                plt.loglog(self.fin,self.fit(x),label='fit ('+tag+')')
+                plt.legend()
+                plt.show() 
+
+    def psd_fn(self, x):
+        # returns the psd function defined earlier           
+
+        if self.fit is not None:
+            xx=x.copy()
+            if self.fit_logx:
+                xx=np.log(x)
+            if self.fit_scale=='linear':
+                dm = np.abs(self.fit(xx))            
+                if self.Sinit is not None:
+                    dm = dm*self.Sinit(x)
+            elif self.fit_scale=='log':
+                logdm = self.fit(xx)
+                zero=self.fit_zero
+                if self.Sinit is not None:
+                    logdm*=(self.logSinit(np.log(x))-np.log(zero))
+                dm=zero*np.exp(logdm)
+            elif self.fit_scale=='logratio':
+                logdm = self.fit(xx)
+                if self.Sinit is not None:
+                    logdm+=self.logSinit(np.log(x))
+                dm=np.exp(logdm)
+            if self.offset_log_fit:
+                dm=dm*np.exp(np.euler_gamma)
+        else:
+            dm = self.Sinit(x)
+
+
+        return dm
